# Cherry-pick: MM-50435: Improve data retention queries on MySQL (#22304)