--- a/store/sqlstore/thread_store.go
+++ b/store/sqlstore/thread_store.go
@@ -900,16 +900,33 @@ func (s *SqlThreadStore) PermanentDeleteBatchThreadMembershipsForRetentionPolici
 
 // DeleteOrphanedRows removes orphaned rows from Threads and ThreadMemberships
 func (s *SqlThreadStore) DeleteOrphanedRows(limit int) (deleted int64, err error) {
+	var threadsQuery string
 	// We need the extra level of nesting to deal with MySQL's locking
-	const threadsQuery = `
+	if s.DriverName() == model.DatabaseDriverMysql {
+		// MySQL fails to do a proper antijoin if the selecting column
+		// and the joining column are different. In that case, doing a subquery
+		// leads to a faster plan because MySQL materializes the sub-query
+		// and does a covering index scan on Threads table. More details on the PR with
+		// this commit.
+		threadsQuery = `
 	DELETE FROM Threads WHERE PostId IN (
 		SELECT * FROM (
+			SELECT Threads.PostId FROM Threads
+			WHERE Threads.ChannelId NOT IN (SELECT Id FROM Channels USE INDEX(PRIMARY))
+			LIMIT ?
+		) AS A
+	)`
+	} else {
+		threadsQuery = `
+	DELETE FROM Threads WHERE PostId IN (
+        SELECT * FROM (
 			SELECT Threads.PostId FROM Threads
 			LEFT JOIN Channels ON Threads.ChannelId = Channels.Id
 			WHERE Channels.Id IS NULL
 			LIMIT ?
 		) AS A
 	)`
+	}
 	// We only delete a thread membership if the entire thread no longer exists,
 	// not if the root post has been deleted
 	const threadMembershipsQuery = `